Rename `cggmp21` to `protocols`
