+++ src/dev.rs
@@ -16,8 +16,8 @@ use ::{
 };
 
 use crate::{
-    cggmp21::{convert_uint, upcast_uint, SchemeParams},
     paillier::PaillierParams,
+    protocols::{convert_uint, upcast_uint, SchemeParams},
 };
 
 #[cfg(feature = "bip32")]
