+++ src/protocols/aux_gen.rs
