+++ src/protocols/aux_gen_tests.rs
